+++ webapp/frontend/src/app/data/mock/device/details/sdd.ts
@@ -15,7 +15,7 @@ export const sdd = {
             'rotational_speed': 0,
             'capacity': 8001563222016,
             'form_factor': '',
-            'smart_support': false,
+            'smart_support': { 'available': false },
             'device_protocol': 'SCSI',
             'device_type': '',
             'label': '',