+++ webapp/frontend/src/app/layout/common/dashboard-device/dashboard-device.component.ts
@@ -74,6 +74,10 @@ export class DashboardDeviceComponent implements OnInit {
         }
     }
 
+    deviceDashboardTitle(): string {
+        return DeviceTitlePipe.deviceDashboardTitle(this.deviceSummary.device);
+    }
+
     openArchiveDialog(): void {
         if(this.deviceSummary.device.archived){
             this._archiveService.unarchiveDevice(this.deviceSummary.device.device_id).subscribe((result) => {