+++ webapp/frontend/src/app/modules/btrfs-filesystem-detail/btrfs-filesystem-detail.routing.ts
@@ -1,0 +1,13 @@
+import { Route } from '@angular/router';
+import { BtrfsFilesystemDetailComponent } from 'app/modules/btrfs-filesystem-detail/btrfs-filesystem-detail.component';
+import { BtrfsFilesystemDetailResolver } from './btrfs-filesystem-detail.resolvers';
+
+export const btrfsFilesystemDetailRoutes: Route[] = [
+    {
+        path: '',
+        component: BtrfsFilesystemDetailComponent,
+        resolve: {
+            filesystem: BtrfsFilesystemDetailResolver
+        }
+    }
+];